add new logic and fix bug

+++ signer/provider/privoder.go
@@ -1,5 +1,5 @@
 package provider
 
 type PriceProvider interface {
-	PriceNow() (float32, error)
+	PriceNow() (float64, error)
 }
